requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
billiard==4.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
celery==5.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
environs==11.0.0; python_version >= '3.8'
flask==3.0.3; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
kombu==5.3.7; python_version >= '3.8'
lxml==5.2.2; python_version >= '3.6'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.3; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
pymupdf==1.24.5; python_version >= '3.8'
pymupdfb==1.24.3; python_version >= '3.8'
pytesseract==0.3.10; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==1.1.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
redis==5.0.5; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.30; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
werkzeug==3.0.3; python_version >= '3.8'
